--- src/main/java/inha/cumulonimbus_cloud/api/controller/auth/dto/response/GetKakaoRes.java
@@ -1,20 +1,0 @@
-package inha.cumulonimbus_cloud.api.controller.auth.dto.response;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@Builder
-@AllArgsConstructor
-public class GetKakaoRes {
-
-    String name;
-    String id;
-
-}
-
-
