--- src/main/java/inha/cumulonimbus_cloud/api/controller/chat/dto/response/PostChatRes.java
@@ -1,15 +1,0 @@
-package inha.cumulonimbus_cloud.api.controller.chat.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@Builder
-@AllArgsConstructor
-public class PostChatRes {
-
-        private String answer;
-}
