+++ src/main/java/inha/cumulonimbus_cloud/auth/api/controller/dto/response/GetKakaoRes.java
@@ -1,0 +1,19 @@
+package inha.cumulonimbus_cloud.auth.api.controller.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@Builder
+@AllArgsConstructor
+public class GetKakaoRes {
+
+    String name;
+    String id;
+
+}
+
+
