+++ src/main/java/inha/cumulonimbus_cloud/auth/api/service/social/kakao/KakaoLoginService.java
@@ -1,0 +1,10 @@
+package inha.cumulonimbus_cloud.auth.api.service.social.kakao;
+
+
+import inha.cumulonimbus_cloud.auth.api.controller.dto.response.GetKakaoRes;
+
+public interface KakaoLoginService {
+
+    String getAccessToken(String authorizationCode);
+    GetKakaoRes getUserInfo(String accessToken);
+}
