+++ src/main/java/inha/cumulonimbus_cloud/chat/api/controller/dto/response/PostChatRes.java
@@ -1,0 +1,15 @@
+package inha.cumulonimbus_cloud.chat.api.controller.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@Builder
+@AllArgsConstructor
+public class PostChatRes {
+
+        private String answer;
+}
