+++ src/main/java/inha/cumulonimbus_cloud/common/code/status/ErrorStatus.java
@@ -41,6 +41,8 @@ public enum ErrorStatus implements BaseErrorCode {
     INVALID_SIZE_10(HttpStatus.BAD_REQUEST, "PAGE4001", "사이즈는 10이어야 합니다."),
     //json 변환
     JSON_CONVERT_ERROR(HttpStatus.BAD_REQUEST, "JSON4000", "JSON 변환에 실패하였습니다."),
+    //s3 업로드 실패
+    S3_UPLOAD_ERROR(HttpStatus.BAD_REQUEST, "S34000", "S3 업로드에 실패하였습니다."),
 
 
     /**
