--- src/main/java/inha/cumulonimbus_cloud/domain/user/User.java
@@ -1,88 +1,0 @@
-package inha.cumulonimbus_cloud.domain.user;
-
-import inha.cumulonimbus_cloud.common.BaseEntity;
-import inha.cumulonimbus_cloud.domain.user.enums.Role;
-import inha.cumulonimbus_cloud.domain.user.enums.SocialType;
-import jakarta.persistence.*;
-import lombok.*;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@EqualsAndHashCode(callSuper = false)
-@Getter
-@Builder
-@Entity
-@Table(name = "users")
-public class User extends BaseEntity implements UserDetails {
-
-    @Id
-    @Column(name = "user_id", nullable = false, updatable = false)
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-
-    @Column(nullable = false, length = 20)
-    private String username;
-
-    @Column(length = 100)
-    private String password;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, length = 20, name = "social_type")
-    private SocialType socialType;
-
-
-    @Column(length = 20)
-    private String nickname;
-
-    private boolean gender;
-
-    @Column(nullable = false, name = "is_finished")
-    private boolean isFinished = false;
-
-
-    @Column(length = 10)
-    @Enumerated(EnumType.STRING)
-    private Role role;
-
-
-
-    @Override
-    public String getUsername() {
-        return username;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return role.getAuthorities();
-    }
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-
-}
