--- src/main/java/inha/cumulonimbus_cloud/domain/user/enums/Permission.java
@@ -1,20 +1,0 @@
-package inha.cumulonimbus_cloud.domain.user.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public enum Permission {
-
-    ADMIN_READ("admin:read"),
-    ADMIN_UPDATE("admin:update"),
-    ADMIN_CREATE("admin:create"),
-    ADMIN_DELETE("admin:delete"),
-    MANAGER_READ("management:read"),
-    MANAGER_UPDATE("management:update"),
-    MANAGER_CREATE("management:create"),
-    MANAGER_DELETE("management:delete");
-
-    @Getter
-    private final String permission;
-}
