--- src/main/java/inha/cumulonimbus_cloud/domain/user/enums/SocialType.java
@@ -1,18 +1,0 @@
-package inha.cumulonimbus_cloud.domain.user.enums;
-
-public enum SocialType {
-    NONE("none"),
-    GOOGLE("google"),
-    KAKAO("kakao"),
-    NAVER("naver");
-
-    private final String description;
-
-    SocialType(String description) {
-        this.description = description;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-}
