+++ src/main/java/inha/cumulonimbus_cloud/user/api/controller/UserController.java
@@ -1,0 +1,4 @@
+package inha.cumulonimbus_cloud.user.api.controller;
+
+public class UserController {
+}
