+++ src/main/java/inha/cumulonimbus_cloud/user/domain/enums/Permission.java
@@ -1,0 +1,20 @@
+package inha.cumulonimbus_cloud.user.domain.enums;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum Permission {
+
+    ADMIN_READ("admin:read"),
+    ADMIN_UPDATE("admin:update"),
+    ADMIN_CREATE("admin:create"),
+    ADMIN_DELETE("admin:delete"),
+    MANAGER_READ("management:read"),
+    MANAGER_UPDATE("management:update"),
+    MANAGER_CREATE("management:create"),
+    MANAGER_DELETE("management:delete");
+
+    @Getter
+    private final String permission;
+}
