+++ src/main/java/inha/cumulonimbus_cloud/user/domain/enums/SocialType.java
@@ -1,0 +1,18 @@
+package inha.cumulonimbus_cloud.user.domain.enums;
+
+public enum SocialType {
+    NONE("none"),
+    GOOGLE("google"),
+    KAKAO("kakao"),
+    NAVER("naver");
+
+    private final String description;
+
+    SocialType(String description) {
+        this.description = description;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
